+++ tests/rules/test_list.py
@@ -5,15 +5,15 @@ import vakt.rules.list
 
 def test_initialization():
     with pytest.raises(TypeError) as excinfo:
-        vakt.rules.list.InList(1)
+        vakt.rules.list.In(1)
     assert 'Initial data should be of list type' == str(excinfo.value)
 
     with pytest.raises(TypeError) as excinfo:
-        vakt.rules.list.InList([[1, 5], [1, 9]])
+        vakt.rules.list.In([[1, 5], [1, 9]])
     assert 'unhashable' in str(excinfo.value)
 
     with pytest.raises(TypeError) as excinfo:
-        vakt.rules.list.InList([{'a': 90}, {'v': "value"}])
+        vakt.rules.list.In([{'a': 90}, {'v': "value"}])
     assert 'unhashable' in str(excinfo.value)
 
 
@@ -28,11 +28,11 @@ def test_initialization():
     (['a', 'b'], 'c', False),
 ])
 def test_in_list(data, what, result):
-    c = vakt.rules.list.InList(data)
+    c = vakt.rules.list.In(data)
     assert result == c.satisfied(what)
     # test after (de)serialization
-    jsn = vakt.rules.list.InList(data).to_json()
-    c1 = vakt.rules.list.InList.from_json(jsn)
+    jsn = vakt.rules.list.In(data).to_json()
+    c1 = vakt.rules.list.In.from_json(jsn)
     assert result == c1.satisfied(what)
 
 
@@ -47,11 +47,11 @@ def test_in_list(data, what, result):
     (['a', 'b'], 'c', True),
 ])
 def test_not_in_list(data, what, result):
-    c = vakt.rules.list.NotInList(data)
+    c = vakt.rules.list.NotIn(data)
     assert result == c.satisfied(what)
     # test after (de)serialization
-    jsn = vakt.rules.list.NotInList(data).to_json()
-    c1 = vakt.rules.list.NotInList.from_json(jsn)
+    jsn = vakt.rules.list.NotIn(data).to_json()
+    c1 = vakt.rules.list.NotIn.from_json(jsn)
     assert result == c1.satisfied(what)
 
 
@@ -67,11 +67,11 @@ def test_not_in_list(data, what, result):
     (['f', 'a', 'y'], ['f', 'a', 'y', 'u'], False),
 ])
 def test_all_in_list(data, what, result):
-    c = vakt.rules.list.AllInList(data)
+    c = vakt.rules.list.AllIn(data)
     assert result == c.satisfied(what)
     # test after (de)serialization
-    jsn = vakt.rules.list.AllInList(data).to_json()
-    c1 = vakt.rules.list.AllInList.from_json(jsn)
+    jsn = vakt.rules.list.AllIn(data).to_json()
+    c1 = vakt.rules.list.AllIn.from_json(jsn)
     assert result == c1.satisfied(what)
 
 
@@ -83,7 +83,7 @@ def test_all_in_list(data, what, result):
     {'a': 90},
 ])
 def test_all_in_list_satisfied_wrong_arg(val):
-    rule = vakt.rules.list.AllInList(['io', 'util'])
+    rule = vakt.rules.list.AllIn(['io', 'util'])
     with pytest.raises(TypeError) as excinfo:
         rule.satisfied(val)
     assert 'Value should be of list type' == str(excinfo.value)
@@ -107,11 +107,11 @@ def test_all_in_list_satisfied_wrong_arg(val):
     (['f', 'a', 'y'], ['f', 'a', 'y', 'u'], True),
 ])
 def test_all_not_in_list(data, what, result):
-    c = vakt.rules.list.AllNotInList(data)
+    c = vakt.rules.list.AllNotIn(data)
     assert result == c.satisfied(what)
     # test after (de)serialization
-    jsn = vakt.rules.list.AllNotInList(data).to_json()
-    c1 = vakt.rules.list.AllNotInList.from_json(jsn)
+    jsn = vakt.rules.list.AllNotIn(data).to_json()
+    c1 = vakt.rules.list.AllNotIn.from_json(jsn)
     assert result == c1.satisfied(what)
 
 
@@ -123,7 +123,7 @@ def test_all_not_in_list(data, what, result):
     {'a': 90},
 ])
 def test_all_not_in_list_satisfied_wrong_arg(val):
-    rule = vakt.rules.list.AllNotInList([])
+    rule = vakt.rules.list.AllNotIn([])
     with pytest.raises(TypeError) as excinfo:
         rule.satisfied(val)
     assert 'Value should be of list type' == str(excinfo.value)
@@ -149,11 +149,11 @@ def test_all_not_in_list_satisfied_wrong_arg(val):
     (['f', 'a', 'y'], ['f', 'a', 'y', 'u'], True),
 ])
 def test_any_in_list(data, what, result):
-    c = vakt.rules.list.AnyInList(data)
+    c = vakt.rules.list.AnyIn(data)
     assert result == c.satisfied(what)
     # test after (de)serialization
-    jsn = vakt.rules.list.AnyInList(data).to_json()
-    c1 = vakt.rules.list.AnyInList.from_json(jsn)
+    jsn = vakt.rules.list.AnyIn(data).to_json()
+    c1 = vakt.rules.list.AnyIn.from_json(jsn)
     assert result == c1.satisfied(what)
 
 
@@ -165,7 +165,7 @@ def test_any_in_list(data, what, result):
     {'a': 90},
 ])
 def test_any_in_list_satisfied_wrong_arg(val):
-    rule = vakt.rules.list.AnyInList(['oop'])
+    rule = vakt.rules.list.AnyIn(['oop'])
     with pytest.raises(TypeError) as excinfo:
         rule.satisfied(val)
     assert 'Value should be of list type' == str(excinfo.value)
@@ -196,11 +196,11 @@ def test_any_in_list_satisfied_wrong_arg(val):
     (['f', 'a', 'y'], ['f', 'a', 'y', 'u'], True),
 ])
 def test_any_not_in_list(data, what, result):
-    c = vakt.rules.list.AnyNotInList(data)
+    c = vakt.rules.list.AnyNotIn(data)
     assert result == c.satisfied(what)
     # test after (de)serialization
-    jsn = vakt.rules.list.AnyNotInList(data).to_json()
-    c1 = vakt.rules.list.AnyNotInList.from_json(jsn)
+    jsn = vakt.rules.list.AnyNotIn(data).to_json()
+    c1 = vakt.rules.list.AnyNotIn.from_json(jsn)
     assert result == c1.satisfied(what)
 
 
@@ -212,7 +212,7 @@ def test_any_not_in_list(data, what, result):
     {'a': 90},
 ])
 def test_any_not_in_list_satisfied_wrong_arg(val):
-    rule = vakt.rules.list.AnyNotInList(['xyz'])
+    rule = vakt.rules.list.AnyNotIn(['xyz'])
     with pytest.raises(TypeError) as excinfo:
         rule.satisfied(val)
     assert 'Value should be of list type' == str(excinfo.value)